Don't report a MAC failure for a lost session error.
When an HSM session times out an "INVALID_SESSION" error message is returned
in the clear. This currently trips a failure of the MAC check since the code
is expecting a secure response. Add code to check to the response to see if
it's an in-the-clear error before raising it as a MAC failure exception. If
it is an in-the-clear error, raise that instead.