core: refactor `AuthSession::__init__()`
This adds the classmethod `create_session()` which is to be used over
instantiating the AuthSession class directly. This is a breaking
change in order to accommodate using different types of authentication
keys (symmetric vs. asymmetric). Implementations that instantiate
AuthSession directly should switch from using the usual constructor to
`AuthSession::create_session()`.

Co-authored-by: Per Nilsson <per.nilsson@yubico.com>