Don't install docs at top level of wheel
Telling Poetry to install the `COPYING`, `NEWS`, and `README.adoc` files
without qualification causes it to install those files at the very top
level of the wheel, outside the `yubihsm/` directory, and thus to
potentially clash with other packages.

We only need to explicitly install them in the sdist.  The wheel still
ends up with copies of those files in `yubihsm-*.dist-info/`, so no
information is lost.